Optimize stack variables

+++ pkg/glib/v2/gdatetime.go
@@ -1745,5 +1745,5 @@ func NewDateTimeFromGo(t time.Time) *DateTime {
 	// Second offset within a minute in nanoseconds.
 	seconds := (time.Duration(s) * time.Second) + time.Duration(t.Nanosecond())
 
-	return NewDateTime(tz, Y, int(M), D, h, m, seconds.Seconds())
+	return NewDateTime(tz, int32(Y), int32(M), int32(D), int32(h), int32(m), seconds.Seconds())
 }
